+++ vkcs/networking/resource_network.go
@@ -103,7 +103,7 @@ func ResourceNetworkingNetwork() *schema.Resource {
 				ForceNew:         true,
 				Computed:         true,
 				ValidateDiagFunc: ValidateSDN(),
-				Description:      "SDN to use for this resource. Must be one of following: \"neutron\", \"sprut\". Default value is \"neutron\".",
+				Description:      "SDN to use for this resource. Must be one of following: \"neutron\", \"sprut\". Default value is project's default SDN.",
 			},
 
 			"vkcs_services_access": {
@@ -208,7 +208,7 @@ func resourceNetworkingNetworkRead(ctx context.Context, d *schema.ResourceData, 
 	d.Set("port_security_enabled", network.PortSecurityEnabled)
 	d.Set("region", util.GetRegion(d, config))
 	d.Set("private_dns_domain", network.PrivateDNSDomain)
-	d.Set("sdn", GetSDN(d))
+	d.Set("sdn", network.SDN)
 	d.Set("vkcs_services_access", network.ServicesAccess)
 
 	NetworkingReadAttributesTags(d, network.Tags)